Fixes to convergence plot rendering. Also, updates and improvements to documentation.

+++ seekr2/modules/common_base.py
@@ -33,7 +33,7 @@ def strBool(bool_str):
         raise Exception(
             "argument for strBool must be string either 'True' or 'False'.")
 
-def order_files_numerically(file_list, func=int, use_basename=False):
+def order_files_numerically(file_list, func=float, use_basename=False):
     """
     If there is a list of files, order them numerically, not
     alphabetically and return the sorted list of files. Note that
